Make the signed apk script runnable from outside the git root

+++ generate-signed-apk.py
@@ -9,9 +9,23 @@ def chdir_to_script_dir():
     os.chdir(os.path.dirname(__file__))
 
 
+# We're using a bunch of paths relative to the script dir, so
+# make sure to use the script dir as the cwd
+call_path = os.getcwd()
+chdir_to_script_dir()
 
 gradle_properties  = os.path.abspath("./android/gradle.properties")
 key_store = os.path.abspath("./android/app/safe-psyc-release-privkey.keystore")
+
+rel_apk_path = rel_to_call_path("./android/app/build/outputs/apk/release/app-release.apk")
+abs_apk_path = os.path.abspath(rel_apk_path)
+
+
+class BuildException(Exception):
+    pass
+class SecretsException(Exception):
+    pass
+
 
 def install_secrets():
     print "Installing secrets..."
@@ -20,6 +34,7 @@ def install_secrets():
 
 def install_gradle_properties():
     gradle_secrets_file = os.path.abspath("./SECRETS/release-gradle-properties")
+    ensure_secret_exists(gradle_secrets_file)
 
     with open(gradle_secrets_file, "r") as f:
         gradle_secrets = f.read()
@@ -29,8 +44,17 @@ def install_gradle_properties():
 
     print "  * gradle.properties secrets installed"
 
+def ensure_secret_exists(path):
+    if not os.path.isfile(path):
+        print
+        print "unable to find {}, perhaps you haven't decrypted the secrets?".format(path)
+        print_decrypt_help()
+        raise SecretsException(path)
+
 def install_key_store():
     key_store_file = os.path.abspath("./SECRETS/safe-psyc-release-privkey.keystore")
+    ensure_secret_exists(key_store_file)
+
     shutil.copyfile(key_store_file, key_store)
     print "  * keystore installed"
 
@@ -55,8 +79,9 @@ def remove_gradle_secrets():
     print "  * gradle.properties secrets removed"
 
 def remove_key_store():
-    os.remove(key_store)
-    print "  * keystore removed"
+    if os.path.isfile(key_store):
+        os.remove(key_store)
+        print "  * keystore removed"
 
 
 def build_apk():
@@ -72,11 +97,11 @@ def build_apk():
         print
     finally:
         os.chdir('..')
-class BuildException(Exception):
-    pass
 
-rel_apk_path = "./android/app/build/outputs/apk/release/app-release.apk"
-abs_apk_path = os.path.abspath(rel_apk_path)
+
+def rel_to_call_path(path):
+    return "./{}".format(os.path.relpath(path, call_path))
+
 def print_apk_path():
     print "The APK can be found at {}".format(abs_apk_path)
 
@@ -89,6 +114,7 @@ def copy_apk_path_to_clipboard():
 
 def which(program):
     # taken from https://stackoverflow.com/a/377028
+    # could probably be replaced by os.system("command {}".format(program))
     def is_exe(fpath):
         return os.path.isfile(fpath) and os.access(fpath, os.X_OK)
 
@@ -109,6 +135,7 @@ def print_install_help():
     print "adb install -r {}".format(rel_apk_path)
 
 def print_testfairy_instructions():
+    testfairy_script = rel_to_call_path('./testfairy-uploader.sh')
     api_key_file = "./SECRETS/testfairy-api-key"
     api_key_file_is_readable = os.path.isfile(api_key_file) and os.access(api_key_file, os.R_OK)
 
@@ -117,29 +144,30 @@ def print_testfairy_instructions():
             testfairy_api_key = f.read().strip()
 
             print "You can upload it to testfairy by running"
-            print "./testfairy-uploader.sh {} {}".format(testfairy_api_key, rel_apk_path)
+            print "{} {} {}".format(testfairy_script, testfairy_api_key, rel_apk_path)
     else:
         print "You can upload it to testfairy by running"
-        print "./testfairy-uploader.sh testfairy-api-key {}".format(testfairy_api_key, rel_apk_path)
+        print "{} testfairy-api-key {}".format(testfairy_script, rel_apk_path)
         print "The API key can be found in {}, but you don't seem to have decrypted the secrets".format(api_key_file)
-        print "You can decrypt the secrets by running"
-        print "./decrypt-folder ./SECRETS.tar.gz"
+        print_decrypt_help()
+
+def print_decrypt_help():
+    print "You can decrypt the secrets by running"
+    print "{} {}".format(
+        rel_to_call_path("./decrypt-folder"),
+        rel_to_call_path("./SECRETS.tar.gpg"),
+    )
 
 def print_secrets_warning():
     print "Please make sure there is no .keystore file in {}".format(dirname_relative(key_store))
     print "and no RELEASE_ keys in {}".format(abs_to_rel(gradle_properties))
 
 def abs_to_rel(abspath):
-    return "./{}".format(os.path.relpath(gradle_properties, os.getcwd()))
+    return "./{}".format(os.path.relpath(gradle_properties, call_path))
 
 def dirname_relative(abspath):
     relpath = abs_to_rel(abspath)
     return "{}/".format(os.path.dirname(relpath))
-
-
-# We're using a bunch of paths relative to the script dir, so
-# make sure to use the script dir as the cwd
-chdir_to_script_dir()
 
 try:
     install_secrets()
@@ -158,6 +186,8 @@ try:
     print
     print_testfairy_instructions()
     print
+except SecretsException as e:
+    pass
 except BuildException as e:
     pass
 
@@ -167,5 +197,4 @@ finally:
     print
     print_secrets_warning()
     print
-
